+++ ddth-akka-core/src/test/java/com/github/ddth/akka/qnd/QndMultiNodeTickFanOutPubSub3.java
@@ -3,6 +3,7 @@ package com.github.ddth.akka.qnd;
 import java.util.Date;
 import java.util.Random;
 import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicLongArray;
 
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
@@ -17,6 +18,7 @@ import com.github.ddth.akka.scheduling.tickfanout.MultiNodePubSubBasedTickFanOut
 import com.github.ddth.dlock.IDLock;
 import com.github.ddth.dlock.impl.redis.RedisDLockFactory;
 import com.github.ddth.pubsub.impl.universal.idint.UniversalRedisPubSubHub;
+import com.google.common.util.concurrent.AtomicDoubleArray;
 
 import akka.actor.ActorRef;
 import akka.actor.ActorSystem;
@@ -32,44 +34,53 @@ public class QndMultiNodeTickFanOutPubSub3 {
         System.setProperty("org.slf4j.simpleLogger.showShortLogName", "false");
     }
 
-    static Logger LOGGER = LoggerFactory.getLogger("QndMultiNodeTickFanOutPubSub2");
+    static Logger LOGGER = LoggerFactory.getLogger(QndMultiNodeTickFanOutPubSub3.class);
     static Random RAND = new Random(System.currentTimeMillis());
+    static int NUM_WORKERS = 4;
+    static AtomicLongArray COUNTER_HIT = new AtomicLongArray(NUM_WORKERS);
+    static AtomicLong COUNTER = new AtomicLong();
+    static AtomicLong LASTTICK = new AtomicLong(0);
 
-    @Scheduling(value = "*/3 * *", getWorkerCoordinationPolicy = WorkerCoordinationPolicy.GLOBAL_SINGLETON)
+    @Scheduling(value = "*/5 * *", workerCoordinationPolicy = WorkerCoordinationPolicy.GLOBAL_SINGLETON)
     public static class GlobalSingletonWorker extends BaseWorker {
 
-        public GlobalSingletonWorker(IDLock dlock, Long dlockTimeMs) {
-            super(dlock, dlockTimeMs);
-        }
+        private int workerId;
 
-        private AtomicLong COUNTER_TASK = new AtomicLong(0), COUNTER_BUSY = new AtomicLong(0);
+        public GlobalSingletonWorker(IDLock dlock, int workerId) {
+            super(dlock);
+            this.workerId = workerId;
+        }
 
         /**
          * {@inheritDoc}
          */
         @Override
         protected void logBusy(TickMessage tick, boolean isGlobal) {
-            COUNTER_BUSY.incrementAndGet();
+            COUNTER.incrementAndGet();
         }
 
         @Override
         protected void doJob(String dlockId, TickMessage tick) throws InterruptedException {
             Date now = new Date();
             try {
-                COUNTER_TASK.incrementAndGet();
-                // LOGGER.info("{" + self().path() + "}: " + tick.getId() + " /
-                // "
-                // + DateFormatUtils.toString(now, DateFormatUtils.DF_ISO8601) +
-                // " / "
-                // + DateFormatUtils.toString(tick.getTimestamp(),
-                // DateFormatUtils.DF_ISO8601)
-                // + " / " + (now.getTime() - tick.getTimestamp().getTime()));
-                // int sleepMs = 1400 + RAND.nextInt(1000);
-                // Thread.sleep(sleepMs);
-                long numTask = COUNTER_TASK.get();
-                long numBusy = COUNTER_BUSY.get();
-                LOGGER.info("\t{" + self().path() + "} " + numTask + " / " + numBusy + " / "
-                        + (numTask * 100.0 / (numTask + numBusy)));
+                long delta = tick.getTimestamp().getTime() - LASTTICK.get();
+                boolean valid = LASTTICK.get() == 0 || (4500 <= delta && delta <= 5500);
+
+                long numTotal = COUNTER.incrementAndGet();
+                COUNTER_HIT.incrementAndGet(workerId);
+                AtomicDoubleArray rate = new AtomicDoubleArray(NUM_WORKERS);
+                for (int i = 0; i < NUM_WORKERS; i++) {
+                    double d = COUNTER_HIT.get(i) * 100.0 / numTotal;
+                    rate.set(i, Math.round(d * 10.0) / 10.0);
+                }
+
+                LOGGER.info(self().path().name() + ": " + tick.getTimestampStr("HH:mm:ss")
+                        + (valid ? " [Y]" : " [N]") + " / Delay: "
+                        + (now.getTime() - tick.getTimestamp().getTime()) + "ms / Counter: "
+                        + COUNTER_HIT + " / Rate: " + rate);
+
+                LASTTICK.set(tick.getTimestamp().getTime());
+                Thread.sleep(RAND.nextInt(3000));
             } finally {
                 if (!StringUtils.isBlank(dlockId)
                         && System.currentTimeMillis() - now.getTime() > 1000) {
@@ -94,16 +105,14 @@ public class QndMultiNodeTickFanOutPubSub3 {
                 try {
                     System.out.println("Actor system: " + actorSystem);
 
-                    for (int i = 0; i < 4; i++) {
+                    for (int i = 0; i < NUM_WORKERS; i++) {
                         IDLock dlock = dlockFactory.createLock("worker");
                         System.out.println("Actor: " + actorSystem.actorOf(
-                                Props.create(GlobalSingletonWorker.class, dlock, 5000L),
-                                "worker" + i));
+                                Props.create(GlobalSingletonWorker.class, dlock, i), "worker" + i));
                     }
 
                     IDLock dlock = dlockFactory.createLock("demo");
                     System.out.println("DLock: " + dlock);
-
                     ActorRef tickFanOut = MultiNodePubSubBasedTickFanOutActor
                             .newInstance(actorSystem, dlock, pubSub, "pubSubChannel");
                     System.out.println(tickFanOut);